UPGRADE to openstreetmap-carto v4.21.1
Upgrading from stylesheet version 4.20.0 to 4.21.1 includes changes
to the shapefiles used (water polygons INSTEAD of land polygons),
updated indexes, and fixes/updates to feature rendering.

The cookbook recipe has been updated to MOVE shapefile definitions
into the attributes. A handful of resources will run to do the
automatic upgrade from v4.20.0 to v4.21.1 if (and only if) the
attributes have been updated to reference "v4.21.1" or newer. It
should still be safe to run the recipe with "v4.20.0", although
the land shapefiles will NOT be downloaded anymore. If you still
have the land shapefiles installed from a previous version of the
cookbook, then it is safe.

For a FRESH deployment on a new node, DO NOT use "v4.20.0" as the
land shapefiles will not be downloaded, causing rendering errors.
New nodes MUST use "v4.21.1" to use the newer water polygons
instead. Versions above "v4.21.1" MAY work, but it would be better
to wait for an update to this cookbook that officially supports
newer versions.